refactor(configeditor): split area browser into focused files
Split update_v3net_area_browser.go (490 lines) into three files
to comply with the 300-line guideline:
- v3net_area_browser_cmds.go (93 lines) — HTTP commands and message types
- v3net_area_browser_helpers.go (194 lines) — message handlers, exit/save, MsgArea creation
- update_v3net_area_browser.go (225 lines) — key event handlers

Also fixes: use bytes.NewReader instead of strings.NewReader(string(data)),
extract areaBrowserListVisible constant, add safety comment on
defaultLocalBoardName ASCII assumption.
