+++ internal/configeditor/update_v3net_area_browser.go
@@ -1,179 +1,11 @@
 package configeditor
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"strings"
-	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
-
-	"github.com/ViSiON-3/vision-3-bbs/internal/message"
-	"github.com/ViSiON-3/vision-3-bbs/internal/v3net/protocol"
 )
-
-// fetchNALMsg is the result of fetching the NAL from a hub.
-type fetchNALMsg struct {
-	areas []protocol.Area
-	err   error
-}
-
-// subscribeAreasMsg is the result of a subscribe call.
-type subscribeAreasMsg struct {
-	statuses []protocol.AreaSubscriptionStatus
-	err      error
-}
-
-// fetchHubNAL returns a tea.Cmd that GETs /v3net/v1/{network}/nal from the hub.
-// This endpoint is public (no auth required).
-func fetchHubNAL(hubURL, network string) tea.Cmd {
-	return func() tea.Msg {
-		client := &http.Client{Timeout: 10 * time.Second}
-		url := strings.TrimRight(hubURL, "/") + "/v3net/v1/" + network + "/nal"
-		resp, err := client.Get(url)
-		if err != nil {
-			return fetchNALMsg{err: err}
-		}
-		defer resp.Body.Close()
-		if resp.StatusCode != http.StatusOK {
-			return fetchNALMsg{err: fmt.Errorf("hub returned status %d", resp.StatusCode)}
-		}
-		var nal protocol.NAL
-		if err := json.NewDecoder(resp.Body).Decode(&nal); err != nil {
-			return fetchNALMsg{err: fmt.Errorf("decode NAL: %w", err)}
-		}
-		return fetchNALMsg{areas: nal.Areas}
-	}
-}
-
-// subscribeToAreas returns a tea.Cmd that POSTs /v3net/v1/subscribe with area tags.
-// The subscribe endpoint is unauthenticated (bootstrap step). The keystore is
-// only needed to populate node_id and pubkey_b64 in the request body.
-func subscribeToAreas(hubURL, network string, areaTags []string,
-	nodeID, pubKeyB64, bbsName, bbsHost string) tea.Cmd {
-	return func() tea.Msg {
-		req := protocol.SubscribeRequest{
-			Network:   network,
-			NodeID:    nodeID,
-			PubKeyB64: pubKeyB64,
-			BBSName:   bbsName,
-			BBSHost:   bbsHost,
-			AreaTags:  areaTags,
-		}
-		data, err := json.Marshal(req)
-		if err != nil {
-			return subscribeAreasMsg{err: fmt.Errorf("marshal subscribe: %w", err)}
-		}
-
-		client := &http.Client{Timeout: 10 * time.Second}
-		url := strings.TrimRight(hubURL, "/") + "/v3net/v1/subscribe"
-		httpReq, err := http.NewRequest("POST", url, strings.NewReader(string(data)))
-		if err != nil {
-			return subscribeAreasMsg{err: err}
-		}
-		httpReq.Header.Set("Content-Type", "application/json")
-
-		resp, err := client.Do(httpReq)
-		if err != nil {
-			return subscribeAreasMsg{err: err}
-		}
-		defer resp.Body.Close()
-
-		if resp.StatusCode != http.StatusOK {
-			return subscribeAreasMsg{err: fmt.Errorf("subscribe returned status %d", resp.StatusCode)}
-		}
-
-		var sr protocol.SubscribeWithAreasResponse
-		if err := json.NewDecoder(resp.Body).Decode(&sr); err != nil {
-			return subscribeAreasMsg{err: fmt.Errorf("decode subscribe response: %w", err)}
-		}
-		return subscribeAreasMsg{statuses: sr.Areas}
-	}
-}
-
-// defaultLocalBoardName generates a default local board name from the network
-// and area display name. e.g. "felonynet" + "General" → "FelonyNet General".
-func defaultLocalBoardName(network, areaName string) string {
-	if network == "" || areaName == "" {
-		return areaName
-	}
-	titled := strings.ToUpper(network[:1]) + network[1:]
-	return titled + " " + areaName
-}
-
-// handleFetchNALMsg processes the NAL fetch result.
-func (m Model) handleFetchNALMsg(msg fetchNALMsg) (tea.Model, tea.Cmd) {
-	m.areaBrowserLoading = false
-	if msg.err != nil {
-		m.areaBrowserError = fmt.Sprintf("Could not fetch areas: %v", msg.err)
-		return m, nil
-	}
-
-	subscribed := make(map[string]bool)
-	localNames := make(map[string]string)
-
-	if m.wizard != nil && m.wizard.flow == "leaf" {
-		for _, sa := range m.wizard.selectedAreas {
-			if sa.Subscribed {
-				subscribed[sa.Tag] = true
-				localNames[sa.Tag] = sa.LocalBoard
-			}
-		}
-	} else if m.areaBrowserReturn == modeRecordEdit {
-		idx := m.recordEditIdx
-		if idx >= 0 && idx < len(m.configs.V3Net.Leaves) {
-			for _, b := range m.configs.V3Net.Leaves[idx].Boards {
-				subscribed[b] = true
-			}
-			for _, ma := range m.configs.MsgAreas {
-				if ma.AreaType == "v3net" && ma.Network == m.areaBrowserNetwork {
-					localNames[ma.EchoTag] = ma.Name
-				}
-			}
-		}
-	}
-
-	items := make([]areaBrowserItem, 0, len(msg.areas))
-	for _, a := range msg.areas {
-		item := areaBrowserItem{
-			Tag:         a.Tag,
-			Name:        a.Name,
-			Description: a.Description,
-			Subscribed:  subscribed[a.Tag],
-		}
-		if name, ok := localNames[a.Tag]; ok {
-			item.LocalBoard = name
-		} else if item.Subscribed {
-			item.LocalBoard = defaultLocalBoardName(m.areaBrowserNetwork, a.Name)
-		}
-		items = append(items, item)
-	}
-	m.areaBrowserAreas = items
-	m.areaBrowserCursor = 0
-	m.areaBrowserScroll = 0
-	m.areaBrowserError = ""
-	return m, nil
-}
-
-// handleSubscribeAreasMsg processes the subscribe response.
-func (m Model) handleSubscribeAreasMsg(msg subscribeAreasMsg) (tea.Model, tea.Cmd) {
-	if msg.err != nil {
-		m.message = fmt.Sprintf("Subscribe failed: %v", msg.err)
-		return m, nil
-	}
-	statusMap := make(map[string]string)
-	for _, s := range msg.statuses {
-		statusMap[s.Tag] = strings.ToUpper(s.Status)
-	}
-	for i := range m.areaBrowserAreas {
-		if st, ok := statusMap[m.areaBrowserAreas[i].Tag]; ok {
-			m.areaBrowserAreas[i].Status = st
-		}
-	}
-	m.message = "Subscription updated"
-	return m, nil
-}
 
 // updateV3NetAreaBrowser handles key events in the area browser.
 func (m Model) updateV3NetAreaBrowser(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
@@ -390,101 +222,4 @@ func (m Model) updateAreaBrowserManual(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 		m.textInput, cmd = m.textInput.Update(msg)
 		return m, cmd
 	}
-}
-
-// exitAreaBrowser saves selections and returns to the previous mode.
-func (m Model) exitAreaBrowser() (tea.Model, tea.Cmd) {
-	if m.wizard != nil && m.wizard.flow == "leaf" {
-		m.wizard.selectedAreas = make([]areaBrowserItem, len(m.areaBrowserAreas))
-		copy(m.wizard.selectedAreas, m.areaBrowserAreas)
-	} else if m.areaBrowserReturn == modeRecordEdit {
-		idx := m.recordEditIdx
-		if idx >= 0 && idx < len(m.configs.V3Net.Leaves) {
-			var boards []string
-			for _, a := range m.areaBrowserAreas {
-				if a.Subscribed {
-					boards = append(boards, a.Tag)
-				}
-			}
-			m.configs.V3Net.Leaves[idx].Boards = boards
-			m.createBrowserMessageAreas()
-			m.dirty = true
-			m.saveAll()
-			m.recordFields = m.buildRecordFields()
-		}
-	}
-	m.mode = m.areaBrowserReturn
-	return m, nil
-}
-
-// createBrowserMessageAreas creates MsgArea entries for subscribed areas
-// that don't already exist. Delegates to createBrowserMsgAreaIfNeeded.
-func (m *Model) createBrowserMessageAreas() {
-	for _, a := range m.areaBrowserAreas {
-		if !a.Subscribed {
-			continue
-		}
-		m.createBrowserMsgAreaIfNeeded(a.Tag, a.LocalBoard, m.areaBrowserNetwork)
-	}
-}
-
-// createBrowserMsgAreaIfNeeded creates a single MsgArea if one doesn't already
-// exist with the given EchoTag.
-func (m *Model) createBrowserMsgAreaIfNeeded(tag, name, network string) {
-	for _, ma := range m.configs.MsgAreas {
-		if ma.EchoTag == tag {
-			return
-		}
-	}
-	newID := 1
-	maxPos := 0
-	for _, ma := range m.configs.MsgAreas {
-		if ma.ID >= newID {
-			newID = ma.ID + 1
-		}
-		if ma.Position > maxPos {
-			maxPos = ma.Position
-		}
-	}
-	m.configs.MsgAreas = append(m.configs.MsgAreas, message.MessageArea{
-		ID:       newID,
-		Position: maxPos + 1,
-		Tag:      tag,
-		Name:     name,
-		AreaType: "v3net",
-		Network:  network,
-		EchoTag:  tag,
-		AutoJoin: true,
-		ACSRead:  "s10",
-		ACSWrite: "s20",
-		BasePath: "msgbases/" + tag,
-	})
-}
-
-// clampAreaBrowserScroll ensures the cursor is visible in the 10-row window.
-func (m *Model) clampAreaBrowserScroll() {
-	visible := 10
-	if m.areaBrowserCursor < m.areaBrowserScroll {
-		m.areaBrowserScroll = m.areaBrowserCursor
-	}
-	if m.areaBrowserCursor >= m.areaBrowserScroll+visible {
-		m.areaBrowserScroll = m.areaBrowserCursor - visible + 1
-	}
-}
-
-// enterAreaBrowser initializes the area browser and starts the NAL fetch.
-func (m Model) enterAreaBrowser(hubURL, network string, returnMode editorMode) (tea.Model, tea.Cmd) {
-	m.areaBrowserHub = hubURL
-	m.areaBrowserNetwork = network
-	m.areaBrowserAreas = nil
-	m.areaBrowserCursor = 0
-	m.areaBrowserScroll = 0
-	m.areaBrowserLoading = true
-	m.areaBrowserError = ""
-	m.areaBrowserManual = false
-	m.areaBrowserEditing = false
-	m.areaBrowserReturn = returnMode
-	m.message = ""
-	m.mode = modeV3NetAreaBrowser
-	return m, fetchHubNAL(hubURL, network)
 }
